HPCC-15130 Reduce jlib dependencies
Move tbb requirement to thorhelper
